feat: Create all events page
