+++ apps/web/src/custom-pages/all-events-page.tsx
@@ -1,0 +1,31 @@
+import { BackButton } from "../components/back-button";
+import EventCard from "../components/event-card";
+import { fetchEvents } from "../lib/api/external/ilmomasiina";
+import { getScopedI18n } from "../locales/server";
+
+export default async function Page() {
+  const t = await getScopedI18n("ilmomasiina.all-events");
+  const ta = await getScopedI18n("action");
+  const since = new Date();
+  since.setMonth(since.getMonth() - 6);
+  const events = await fetchEvents(since);
+
+  return (
+    <main
+      id="main"
+      className="relative mb-8 flex flex-col items-center gap-2 md:gap-6"
+    >
+      <div className="relative m-auto flex max-w-full flex-col gap-8 p-4 md:p-6">
+        <div className="max-w-4xl space-y-4 md:my-8 md:space-y-8">
+          <BackButton>{ta("Back")}</BackButton>
+          <h1 className="font-mono text-4xl">{t("Kaikki tapahtumat")}</h1>
+          <ul className="space-y-8">
+            {events.data
+              ?.reverse()
+              .map((event) => <EventCard event={event} key={event.id} />)}
+          </ul>
+        </div>
+      </div>
+    </main>
+  );
+}
