+++ apps/web/src/locales/en.ts
@@ -123,6 +123,8 @@ const en = {
   "ilmomasiina.Avoin kiintiö": "Open quota",
   "ilmomasiina.Jonossa": "In queue",
   "ilmomasiina.path.events": "events",
+  "ilmomasiina.path.all-events": "all-events",
+  "ilmomasiina.all-events.Kaikki tapahtumat": "All events",
   "ilmomasiina.status.Ei ilmoittautuneita vielä": "No sign ups yet.",
   "ilmomasiina.status.Ilmoittautuminen alkaa": "Sign ups start on {startDate}",
   "ilmomasiina.status.Ilmoittautuminen auki":
@@ -137,6 +139,7 @@ const en = {
   "ilmomasiina.Tilaa kalenteri": "Subscribe to calendar",
   "ilmomasiina.Kopioidaan leikepöydälle": "Copying to clipboard",
   "ilmomasiina.Kopioitu leikepöydälle": "Copied to clipboard",
+  "ilmomasiina.Selaa vanhoja tapahtumia": "Browse old events",
   "not-found.Etusivulle": "To front page",
   "not-found.Tapahtumalistaukseen": "To event list",
   "not-found.Sivua ei löytynyt": "Page not found",
