+++ apps/web/src/locales/fi.ts
@@ -123,6 +123,8 @@ const fi = {
   "ilmomasiina.Avoin kiintiö": "Avoin kiintiö",
   "ilmomasiina.Jonossa": "Jonossa",
   "ilmomasiina.path.events": "tapahtumat",
+  "ilmomasiina.path.all-events": "kaikki-tapahtumat",
+  "ilmomasiina.all-events.Kaikki tapahtumat": "Kaikki tapahtumat",
   "ilmomasiina.status.Ei ilmoittautuneita vielä": "Ei ilmoittautuneita vielä.",
   "ilmomasiina.status.Ilmoittautuminen alkaa":
     "Ilmoittautuminen alkaa {startDate}",
@@ -139,6 +141,7 @@ const fi = {
   "ilmomasiina.Tilaa kalenteri": "Tilaa kalenteri",
   "ilmomasiina.Kopioidaan leikepöydälle": "Kopioidaan leikepöydälle",
   "ilmomasiina.Kopioitu leikepöydälle": "Kopioitu leikepöydälle",
+  "ilmomasiina.Selaa vanhoja tapahtumia": "Selaa vanhoja tapahtumia",
   "not-found.Etusivulle": "Etusivulle",
   "not-found.Tapahtumalistaukseen": "Tapahtumalistaukseen",
   "not-found.Sivua ei löytynyt": "Sivua ei löytynyt",
